Allow manual placement of ad-hoc filters
The postgres SQL parser that this plugin uses does not support ClickHouse
specific syntax (e.g. `quantile(.50)(foo)` or `INTERVAL 1 day`). Further,
the plugin's logic for detecting the target table from the query can't
handle CTEs. As a result, some queries will just not apply ad-hoc filters.
In the former case, there's at least a SQL error in the console, but the
latter case is completely silent.

This change allows users to manually place the ad-hoc filters in the
SETTINGS clause of their queries using a new `$__adHocFilters` macro.
This bypasses all of the automatic table detection logic, including
the SQL parsing.